Implemented XMLInput (first version), other stuff
-Implemented the first version of XMLInput. Everything compiles fine,
and it so far seems to work fine. Testing is needed, of course, but it
should be about ready for regular operation! Note, however, that one
function is needed from MecanumDrive in order for it to operate (which
should be simple enough). This class is, of course, open to change in
the future.
-There is now a sample XML config file added to the project. While the
"IndMotors" are actually crap, the drivebase bit should be operational,
provided that that function is added to MecanumDrive.
-Added additional code from FIRST - DriverStationLCD. It extends the
SensorBase class (or something), and allows for user messages to be put
on the dashboard without some of the problems with SmartDashboard's way
of doing things.
-Added a small text file with the copyright notices from each of our
used third party libraries in it. This means that the original PugiXML
Note is no longer needed - the new "Lib Note.txt" contains everything
now.