Possibly fixed the dead USB camera output issue
When the switch button is pressed, the camera view - which previously worked - stops working. Trying to switch back to the original camera doesn't bring the camera view back. The only thing I can think of is that the StopCamera function is called. This has been disabled, so hopefully dual camera code will work now.