Fixed motor safety timeout bug
Duplicate of similar commit to master. This code was leftover debug code from when we were trying to figure out why Auton driving was slow (it was because we didn't do SD_RetrievePID()). These nasty little safety disabled were probably responsible for a whole host of "why is the robot moving without a command?" issues too.