Switched XMLInput over from using vectors to maps. Might not work, hence the new branch
Maps are a neat thing in the standard template library that accept a key and a value.
They can handle searching for things automatically (reportedly by using a binary search
or something). Not only is this incredibly fast compared to the "iterate EVERYTHING"
method used before, but it's a lot easier to implement - no more annoying for loops
everywhere for the simplest things.